+++ src/test/java/app/gpx_animator/ui/cli/CommandLineIT.java
@@ -61,8 +61,8 @@ final class CommandLineIT {
         assertDone();
 
         final var fileSize = new File(outputFile).length();
-        assertTrue(fileSize > 250_000, "Output file size (%s bytes) too small, check content".formatted(fileSize));
-        assertTrue(fileSize < 350_000, "Output file size (%s bytes) too big, check content".formatted(fileSize));
+        assertTrue(fileSize > 300_000, "Output file size (%s bytes) too small, check content".formatted(fileSize));
+        assertTrue(fileSize < 400_000, "Output file size (%s bytes) too big, check content".formatted(fileSize));
     }
 
     @AfterEach