+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * New configuration option to specify the user agent when downloading map data
 * New speed unit (/500m)
 * Time per distance speed units now display as m:ss
+* GPS status with timeout (OK/LOST)
 
 ---